fix(`ndt_scan_matcher`): rename source files for consistency
