+++ localization/autoware_ndt_scan_matcher/src/ndt_scan_matcher.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ndt_scan_matcher_helper.hpp"
+#include "ndt_scan_matcher.hpp"
 
 #include <autoware/localization_util/util_func.hpp>
