+++ localization/autoware_ndt_scan_matcher/src/ndt_scan_matcher_node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ndt_scan_matcher_helper.hpp"
+#include "ndt_scan_matcher.hpp"
 
 #include <autoware/localization_util/matrix_type.hpp>
 #include <autoware/localization_util/tree_structured_parzen_estimator.hpp>
